package rename and removed reference to Sample app
