+++ Source/MicroVision/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion '26.0.2'
     defaultConfig {
-        applicationId "com.goertek.microvision"
+        applicationId "com.mvis.apps"
         minSdkVersion 19
         targetSdkVersion 25
         versionCode 1
