+++ Source/MicroVision/app/src/main/java/com/mvis/apps/MessageCenter.java
@@ -1,4 +1,4 @@
-package com.goertek.microvision;
+package com.mvis.apps;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -6,7 +6,6 @@ import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.Message;
 import android.util.Log;
-import android.widget.Toast;
 
 import picop.interfaces.def.ALC_Api;
 import picop.interfaces.def.ALC_Api.ALC_Callback;
@@ -34,111 +33,110 @@ import picop.interfaces.def.PicoP_ValueStorageTypeE;
 import picop.interfaces.def.PicoP_VideoCaptureInfo;
 import picop.interfaces.def.PicoP_VideoModeHandleE;
 
-import static com.goertek.microvision.MainActivity.SERIAL_PORT_OPENED;
-import static com.goertek.microvision.MainActivity.connectionInfo;
-import static com.goertek.microvision.MainActivity.mContext;
-import static com.goertek.microvision.MainActivity.mPicoPHandler;
-import static com.goertek.microvision.Utils.MSG_CLOSE_CONNECTION;
-import static com.goertek.microvision.Utils.MSG_FLIP_IMAGE;
-import static com.goertek.microvision.Utils.MSG_GET_ACTIVECAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_GET_ACTIVEOSD;
-import static com.goertek.microvision.Utils.MSG_GET_ASPECTRATIOMODE;
-import static com.goertek.microvision.Utils.MSG_GET_BRIGHTNESS;
-import static com.goertek.microvision.Utils.MSG_GET_CLEARTARGER;
-import static com.goertek.microvision.Utils.MSG_GET_COLORALIGNMENT;
-import static com.goertek.microvision.Utils.MSG_GET_COLORCONVERTER;
-import static com.goertek.microvision.Utils.MSG_GET_COLOR_MODE;
-import static com.goertek.microvision.Utils.MSG_GET_COMMITINPUTCAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_GET_DISPLAYINFO;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWLINE;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWPOINT;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWRECTANGLE;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTESTPATTERN;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTEXT;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTRIANGLE;
-//import static com.goertek.microvision.Utils.MSG_GET_ENUMERATEDEVICES;
-import static com.goertek.microvision.Utils.MSG_GET_EVENTLOG;
-import static com.goertek.microvision.Utils.MSG_GET_FLIPSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_GAMMAVAL;
-import static com.goertek.microvision.Utils.MSG_GET_INPUTCAPTUREMODEINFO;
-import static com.goertek.microvision.Utils.MSG_GET_INPUTVIDEOPROPERTIES;
-import static com.goertek.microvision.Utils.MSG_GET_INPUTVIDEOSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_KEYSTONE;
-import static com.goertek.microvision.Utils.MSG_GET_LOADBITMAPIMAGE;
-import static com.goertek.microvision.Utils.MSG_GET_OSDINFO;
-import static com.goertek.microvision.Utils.MSG_GET_OSDSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_OUTPUTVIDEOSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_PHASE;
-import static com.goertek.microvision.Utils.MSG_GET_RENDER;
-import static com.goertek.microvision.Utils.MSG_RESTOREFACTORYCONFIG;
-import static com.goertek.microvision.Utils.MSG_GET_SPLASHSCREENTIMEOUT;
-import static com.goertek.microvision.Utils.MSG_GET_SYSTEMINFO;
-import static com.goertek.microvision.Utils.MSG_GET_SYSTEMSTATUS;
-import static com.goertek.microvision.Utils.MSG_GET_TEXTBOXINFO;
-//import static com.goertek.microvision.Utils.MSG_GET_UPGRADESOFTWARE;
-import static com.goertek.microvision.Utils.MSG_GET_VIEWPORTDISTORTION;
-import static com.goertek.microvision.Utils.MSG_MODIFY_INPUTCAPTUREMODEINFO;
-import static com.goertek.microvision.Utils.MSG_SET_ACTIVECAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_SET_ACTIVEOSD;
-import static com.goertek.microvision.Utils.MSG_SET_ASPECTRATIOMODE;
-import static com.goertek.microvision.Utils.MSG_SET_BRIGHTNESS;
-import static com.goertek.microvision.Utils.MSG_SET_COLORALIGNMENT;
-import static com.goertek.microvision.Utils.MSG_SET_COLORCONVERTER;
-import static com.goertek.microvision.Utils.MSG_SET_COLOR_MODE;
-import static com.goertek.microvision.Utils.MSG_SET_COMMITINPUTCAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_SET_FLIPSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_GAMMAVAL;
-import static com.goertek.microvision.Utils.MSG_SET_INPUTCAPTUREMODEINFO;
-import static com.goertek.microvision.Utils.MSG_SET_INPUTVIDEOSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_KEYSTONE;
-import static com.goertek.microvision.Utils.MSG_SET_OSDINFO;
-import static com.goertek.microvision.Utils.MSG_SET_OSDSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_OUTPUTVIDEOSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_PHASE;
+import static com.mvis.apps.MainActivity.SERIAL_PORT_OPENED;
+import static com.mvis.apps.MainActivity.connectionInfo;
+import static com.mvis.apps.MainActivity.mPicoPHandler;
+import static com.mvis.apps.Utils.MSG_CLOSE_CONNECTION;
+import static com.mvis.apps.Utils.MSG_FLIP_IMAGE;
+import static com.mvis.apps.Utils.MSG_GET_ACTIVECAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_GET_ACTIVEOSD;
+import static com.mvis.apps.Utils.MSG_GET_ASPECTRATIOMODE;
+import static com.mvis.apps.Utils.MSG_GET_BRIGHTNESS;
+import static com.mvis.apps.Utils.MSG_GET_CLEARTARGER;
+import static com.mvis.apps.Utils.MSG_GET_COLORALIGNMENT;
+import static com.mvis.apps.Utils.MSG_GET_COLORCONVERTER;
+import static com.mvis.apps.Utils.MSG_GET_COLOR_MODE;
+import static com.mvis.apps.Utils.MSG_GET_COMMITINPUTCAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_GET_DISPLAYINFO;
+import static com.mvis.apps.Utils.MSG_GET_DRAWLINE;
+import static com.mvis.apps.Utils.MSG_GET_DRAWPOINT;
+import static com.mvis.apps.Utils.MSG_GET_DRAWRECTANGLE;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTESTPATTERN;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTEXT;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTRIANGLE;
+//import static com.mvis.apps.Utils.MSG_GET_ENUMERATEDEVICES;
+import static com.mvis.apps.Utils.MSG_GET_EVENTLOG;
+import static com.mvis.apps.Utils.MSG_GET_FLIPSTATE;
+import static com.mvis.apps.Utils.MSG_GET_GAMMAVAL;
+import static com.mvis.apps.Utils.MSG_GET_INPUTCAPTUREMODEINFO;
+import static com.mvis.apps.Utils.MSG_GET_INPUTVIDEOPROPERTIES;
+import static com.mvis.apps.Utils.MSG_GET_INPUTVIDEOSTATE;
+import static com.mvis.apps.Utils.MSG_GET_KEYSTONE;
+import static com.mvis.apps.Utils.MSG_GET_LOADBITMAPIMAGE;
+import static com.mvis.apps.Utils.MSG_GET_OSDINFO;
+import static com.mvis.apps.Utils.MSG_GET_OSDSTATE;
+import static com.mvis.apps.Utils.MSG_GET_OUTPUTVIDEOSTATE;
+import static com.mvis.apps.Utils.MSG_GET_PHASE;
+import static com.mvis.apps.Utils.MSG_GET_RENDER;
+import static com.mvis.apps.Utils.MSG_RESTOREFACTORYCONFIG;
+import static com.mvis.apps.Utils.MSG_GET_SPLASHSCREENTIMEOUT;
+import static com.mvis.apps.Utils.MSG_GET_SYSTEMINFO;
+import static com.mvis.apps.Utils.MSG_GET_SYSTEMSTATUS;
+import static com.mvis.apps.Utils.MSG_GET_TEXTBOXINFO;
+//import static com.mvis.apps.Utils.MSG_GET_UPGRADESOFTWARE;
+import static com.mvis.apps.Utils.MSG_GET_VIEWPORTDISTORTION;
+import static com.mvis.apps.Utils.MSG_MODIFY_INPUTCAPTUREMODEINFO;
+import static com.mvis.apps.Utils.MSG_SET_ACTIVECAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_SET_ACTIVEOSD;
+import static com.mvis.apps.Utils.MSG_SET_ASPECTRATIOMODE;
+import static com.mvis.apps.Utils.MSG_SET_BRIGHTNESS;
+import static com.mvis.apps.Utils.MSG_SET_COLORALIGNMENT;
+import static com.mvis.apps.Utils.MSG_SET_COLORCONVERTER;
+import static com.mvis.apps.Utils.MSG_SET_COLOR_MODE;
+import static com.mvis.apps.Utils.MSG_SET_COMMITINPUTCAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_SET_FLIPSTATE;
+import static com.mvis.apps.Utils.MSG_SET_GAMMAVAL;
+import static com.mvis.apps.Utils.MSG_SET_INPUTCAPTUREMODEINFO;
+import static com.mvis.apps.Utils.MSG_SET_INPUTVIDEOSTATE;
+import static com.mvis.apps.Utils.MSG_SET_KEYSTONE;
+import static com.mvis.apps.Utils.MSG_SET_OSDINFO;
+import static com.mvis.apps.Utils.MSG_SET_OSDSTATE;
+import static com.mvis.apps.Utils.MSG_SET_OUTPUTVIDEOSTATE;
+import static com.mvis.apps.Utils.MSG_SET_PHASE;
 
-import static com.goertek.microvision.Utils.MSG_SET_SPLASHSCREENTIMEOUT;
-import static com.goertek.microvision.Utils.MSG_SET_VIEWPORTDISTORTION;
-import static com.goertek.microvision.fragment.ActiveCaptureMode.MSG_ACTIVEACTIVEMODE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.ActiveCaptureMode.activeCapModeHandler;
-import static com.goertek.microvision.fragment.Brightness.MSG_BRIGHTNESS_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Brightness.brightnessHandler;
-import static com.goertek.microvision.fragment.ColorAlignment.MSG_COLORALIGNMENT_RESPONSE_GET;
-import static com.goertek.microvision.fragment.ColorAlignment.coloralignmentHandler;
-import static com.goertek.microvision.fragment.ColorMode.MSG_COLORMODE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.ColorMode.colorModeHandler;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_ACTIVE_OSD_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_ASPECT_RATIO_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_COLOR_CONVERTER_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_COMMIT_INPUT_CAPTURE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_DISPLAY_INFO_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_FLIP_STATE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_OSD_INFO_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_OSD_STATE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_SPLASH_SCREEN_TIMEOUT_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_TEMPERATURE_LOOP_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_TEXTBOX_INFO_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.MSG_FUNCTIONS_VIEWPORT_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Functions.functionsHandler;
-import static com.goertek.microvision.fragment.GammaVal.MSG_GAMMAVAL_RESPONSE_GET;
-import static com.goertek.microvision.fragment.GammaVal.gammavalHandler;
-import static com.goertek.microvision.fragment.GetEventLog.MSG_EVENTLOG_RESPONSE_GET;
-import static com.goertek.microvision.fragment.GetEventLog.eventlogHandler;
-import static com.goertek.microvision.fragment.InputCaptureMode.MSG_INPUTCAPTURE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.InputCaptureMode.inputCaptureHandler;
-import static com.goertek.microvision.fragment.InputVideoProperties.MSG_GET_INPUTVIDEOPROPERTIES_RESPONSE_GET;
-import static com.goertek.microvision.fragment.InputVideoProperties.inputvideopropertiesHandler;
-import static com.goertek.microvision.fragment.InputVideoState.MSG_INPUTVIDEOSTATE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.InputVideoState.inputvideostateHandler;
-import static com.goertek.microvision.fragment.KeyStone.MSG_KEYSTONE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.KeyStone.keystoneHandler;
-import static com.goertek.microvision.fragment.OutputVideoState.MSG_OUTPUTVIDEOSTATE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.OutputVideoState.outputvideostateHandler;
-import static com.goertek.microvision.fragment.Phase.MSG_PHASE_RESPONSE_GET;
-import static com.goertek.microvision.fragment.Phase.phaseHandler;
-import static com.goertek.microvision.fragment.SystemInfo.MSG_GET_SYSTEMINFO_RESPONSE_GET;
-import static com.goertek.microvision.fragment.SystemInfo.systeminfoHandler;
-import static com.goertek.microvision.fragment.SystemStatus.MSG_GET_SYSTEMSTATUS_RESPONSE_GET;
-import static com.goertek.microvision.fragment.SystemStatus.systemstatusHandler;
+import static com.mvis.apps.Utils.MSG_SET_SPLASHSCREENTIMEOUT;
+import static com.mvis.apps.Utils.MSG_SET_VIEWPORTDISTORTION;
+import static com.mvis.apps.fragment.ActiveCaptureMode.MSG_ACTIVEACTIVEMODE_RESPONSE_GET;
+import static com.mvis.apps.fragment.ActiveCaptureMode.activeCapModeHandler;
+import static com.mvis.apps.fragment.Brightness.MSG_BRIGHTNESS_RESPONSE_GET;
+import static com.mvis.apps.fragment.Brightness.brightnessHandler;
+import static com.mvis.apps.fragment.ColorAlignment.MSG_COLORALIGNMENT_RESPONSE_GET;
+import static com.mvis.apps.fragment.ColorAlignment.coloralignmentHandler;
+import static com.mvis.apps.fragment.ColorMode.MSG_COLORMODE_RESPONSE_GET;
+import static com.mvis.apps.fragment.ColorMode.colorModeHandler;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_ACTIVE_OSD_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_ASPECT_RATIO_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_COLOR_CONVERTER_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_COMMIT_INPUT_CAPTURE_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_DISPLAY_INFO_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_FLIP_STATE_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_OSD_INFO_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_OSD_STATE_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_SPLASH_SCREEN_TIMEOUT_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_TEMPERATURE_LOOP_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_TEXTBOX_INFO_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.MSG_FUNCTIONS_VIEWPORT_RESPONSE_GET;
+import static com.mvis.apps.fragment.Functions.functionsHandler;
+import static com.mvis.apps.fragment.GammaVal.MSG_GAMMAVAL_RESPONSE_GET;
+import static com.mvis.apps.fragment.GammaVal.gammavalHandler;
+import static com.mvis.apps.fragment.GetEventLog.MSG_EVENTLOG_RESPONSE_GET;
+import static com.mvis.apps.fragment.GetEventLog.eventlogHandler;
+import static com.mvis.apps.fragment.InputCaptureMode.MSG_INPUTCAPTURE_RESPONSE_GET;
+import static com.mvis.apps.fragment.InputCaptureMode.inputCaptureHandler;
+import static com.mvis.apps.fragment.InputVideoProperties.MSG_GET_INPUTVIDEOPROPERTIES_RESPONSE_GET;
+import static com.mvis.apps.fragment.InputVideoProperties.inputvideopropertiesHandler;
+import static com.mvis.apps.fragment.InputVideoState.MSG_INPUTVIDEOSTATE_RESPONSE_GET;
+import static com.mvis.apps.fragment.InputVideoState.inputvideostateHandler;
+import static com.mvis.apps.fragment.KeyStone.MSG_KEYSTONE_RESPONSE_GET;
+import static com.mvis.apps.fragment.KeyStone.keystoneHandler;
+import static com.mvis.apps.fragment.OutputVideoState.MSG_OUTPUTVIDEOSTATE_RESPONSE_GET;
+import static com.mvis.apps.fragment.OutputVideoState.outputvideostateHandler;
+import static com.mvis.apps.fragment.Phase.MSG_PHASE_RESPONSE_GET;
+import static com.mvis.apps.fragment.Phase.phaseHandler;
+import static com.mvis.apps.fragment.SystemInfo.MSG_GET_SYSTEMINFO_RESPONSE_GET;
+import static com.mvis.apps.fragment.SystemInfo.systeminfoHandler;
+import static com.mvis.apps.fragment.SystemStatus.MSG_GET_SYSTEMSTATUS_RESPONSE_GET;
+import static com.mvis.apps.fragment.SystemStatus.systemstatusHandler;
 import static picop.interfaces.def.PicoP_ColorModeE.eCOLOR_MODE_BRILLIANT;
 import static picop.interfaces.def.PicoP_PpcpUtils.MAX_EVENT_LOG;
