+++ Source/MicroVision/app/src/main/java/com/mvis/apps/fragment/ColorMode.java
@@ -1,4 +1,4 @@
-package com.goertek.microvision.fragment;
+package com.mvis.apps.fragment;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -12,15 +12,13 @@ import android.widget.SeekBar;
 import android.widget.TextView;
 
 
-import com.goertek.microvision.R;
+import com.mvis.apps.R;
 
-import picop.interfaces.def.PicoP_ColorModeE;
-
-import static com.goertek.microvision.MessageCenter.colorMode;
-import static com.goertek.microvision.MessageCenter.messageHandler;
-import static com.goertek.microvision.MessageCenter.resultData;
-import static com.goertek.microvision.Utils.MSG_GET_COLOR_MODE;
-import static com.goertek.microvision.Utils.MSG_SET_COLOR_MODE;
+import static com.mvis.apps.MessageCenter.colorMode;
+import static com.mvis.apps.MessageCenter.messageHandler;
+import static com.mvis.apps.MessageCenter.resultData;
+import static com.mvis.apps.Utils.MSG_GET_COLOR_MODE;
+import static com.mvis.apps.Utils.MSG_SET_COLOR_MODE;
 
 /**
  * Created on 2017/4/7.
