+++ Source/MicroVision/app/src/main/java/com/mvis/apps/fragment/Functions.java
@@ -1,4 +1,4 @@
-package com.goertek.microvision.fragment;
+package com.mvis.apps.fragment;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -10,42 +10,42 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.goertek.microvision.R;
+import com.mvis.apps.R;
 
-import static com.goertek.microvision.MessageCenter.messageHandler;
-import static com.goertek.microvision.MessageCenter.resultData;
-import static com.goertek.microvision.Utils.MSG_GET_ACTIVEOSD;
-import static com.goertek.microvision.Utils.MSG_GET_ASPECTRATIOMODE;
-import static com.goertek.microvision.Utils.MSG_GET_CLEARTARGER;
-import static com.goertek.microvision.Utils.MSG_GET_COLORCONVERTER;
-import static com.goertek.microvision.Utils.MSG_GET_COMMITINPUTCAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_GET_DISPLAYINFO;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWLINE;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWPOINT;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWRECTANGLE;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTESTPATTERN;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTEXT;
-import static com.goertek.microvision.Utils.MSG_GET_DRAWTRIANGLE;
-//import static com.goertek.microvision.Utils.MSG_GET_ENUMERATEDEVICES;
-import static com.goertek.microvision.Utils.MSG_GET_FLIPSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_LOADBITMAPIMAGE;
-import static com.goertek.microvision.Utils.MSG_GET_OSDINFO;
-import static com.goertek.microvision.Utils.MSG_GET_OSDSTATE;
-import static com.goertek.microvision.Utils.MSG_GET_RENDER;
-import static com.goertek.microvision.Utils.MSG_RESTOREFACTORYCONFIG;
-import static com.goertek.microvision.Utils.MSG_GET_SPLASHSCREENTIMEOUT;
-import static com.goertek.microvision.Utils.MSG_GET_TEXTBOXINFO;
-//import static com.goertek.microvision.Utils.MSG_GET_UPGRADESOFTWARE;
-import static com.goertek.microvision.Utils.MSG_GET_VIEWPORTDISTORTION;
-import static com.goertek.microvision.Utils.MSG_SET_ACTIVEOSD;
-import static com.goertek.microvision.Utils.MSG_SET_ASPECTRATIOMODE;
-import static com.goertek.microvision.Utils.MSG_SET_COLORCONVERTER;
-import static com.goertek.microvision.Utils.MSG_SET_COMMITINPUTCAPTUREMODE;
-import static com.goertek.microvision.Utils.MSG_SET_FLIPSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_OSDINFO;
-import static com.goertek.microvision.Utils.MSG_SET_OSDSTATE;
-import static com.goertek.microvision.Utils.MSG_SET_SPLASHSCREENTIMEOUT;
-import static com.goertek.microvision.Utils.MSG_SET_VIEWPORTDISTORTION;
+import static com.mvis.apps.MessageCenter.messageHandler;
+import static com.mvis.apps.MessageCenter.resultData;
+import static com.mvis.apps.Utils.MSG_GET_ACTIVEOSD;
+import static com.mvis.apps.Utils.MSG_GET_ASPECTRATIOMODE;
+import static com.mvis.apps.Utils.MSG_GET_CLEARTARGER;
+import static com.mvis.apps.Utils.MSG_GET_COLORCONVERTER;
+import static com.mvis.apps.Utils.MSG_GET_COMMITINPUTCAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_GET_DISPLAYINFO;
+import static com.mvis.apps.Utils.MSG_GET_DRAWLINE;
+import static com.mvis.apps.Utils.MSG_GET_DRAWPOINT;
+import static com.mvis.apps.Utils.MSG_GET_DRAWRECTANGLE;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTESTPATTERN;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTEXT;
+import static com.mvis.apps.Utils.MSG_GET_DRAWTRIANGLE;
+//import static com.mvis.apps.Utils.MSG_GET_ENUMERATEDEVICES;
+import static com.mvis.apps.Utils.MSG_GET_FLIPSTATE;
+import static com.mvis.apps.Utils.MSG_GET_LOADBITMAPIMAGE;
+import static com.mvis.apps.Utils.MSG_GET_OSDINFO;
+import static com.mvis.apps.Utils.MSG_GET_OSDSTATE;
+import static com.mvis.apps.Utils.MSG_GET_RENDER;
+import static com.mvis.apps.Utils.MSG_RESTOREFACTORYCONFIG;
+import static com.mvis.apps.Utils.MSG_GET_SPLASHSCREENTIMEOUT;
+import static com.mvis.apps.Utils.MSG_GET_TEXTBOXINFO;
+//import static com.mvis.apps.Utils.MSG_GET_UPGRADESOFTWARE;
+import static com.mvis.apps.Utils.MSG_GET_VIEWPORTDISTORTION;
+import static com.mvis.apps.Utils.MSG_SET_ACTIVEOSD;
+import static com.mvis.apps.Utils.MSG_SET_ASPECTRATIOMODE;
+import static com.mvis.apps.Utils.MSG_SET_COLORCONVERTER;
+import static com.mvis.apps.Utils.MSG_SET_COMMITINPUTCAPTUREMODE;
+import static com.mvis.apps.Utils.MSG_SET_FLIPSTATE;
+import static com.mvis.apps.Utils.MSG_SET_OSDINFO;
+import static com.mvis.apps.Utils.MSG_SET_OSDSTATE;
+import static com.mvis.apps.Utils.MSG_SET_SPLASHSCREENTIMEOUT;
+import static com.mvis.apps.Utils.MSG_SET_VIEWPORTDISTORTION;
 
 /**
  * Created on 2017/4/24.
