+++ Source/MicroVision/app/src/main/java/com/mvis/apps/fragment/InputCaptureMode.java
@@ -1,4 +1,4 @@
-package com.goertek.microvision.fragment;
+package com.mvis.apps.fragment;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -16,12 +16,12 @@ import android.widget.PopupMenu;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
-import com.goertek.microvision.R;
+import com.mvis.apps.R;
 
-import static com.goertek.microvision.MessageCenter.inputCaptureModeInfo;
-import static com.goertek.microvision.MessageCenter.messageHandler;
-import static com.goertek.microvision.Utils.MSG_GET_INPUTCAPTUREMODEINFO;
-import static com.goertek.microvision.Utils.MSG_SET_INPUTCAPTUREMODEINFO;
+import static com.mvis.apps.MessageCenter.inputCaptureModeInfo;
+import static com.mvis.apps.MessageCenter.messageHandler;
+import static com.mvis.apps.Utils.MSG_GET_INPUTCAPTUREMODEINFO;
+import static com.mvis.apps.Utils.MSG_SET_INPUTCAPTUREMODEINFO;
 
 /**
  * Created on 2017/4/22.
