+++ Source/MicroVision/app/src/main/java/com/mvis/apps/fragment/SystemInfo.java
@@ -1,4 +1,4 @@
-package com.goertek.microvision.fragment;
+package com.mvis.apps.fragment;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -10,11 +10,11 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.goertek.microvision.R;
+import com.mvis.apps.R;
 
-import static com.goertek.microvision.MessageCenter.messageHandler;
-import static com.goertek.microvision.MessageCenter.systemInfo;
-import static com.goertek.microvision.Utils.MSG_GET_SYSTEMINFO;
+import static com.mvis.apps.MessageCenter.messageHandler;
+import static com.mvis.apps.MessageCenter.systemInfo;
+import static com.mvis.apps.Utils.MSG_GET_SYSTEMINFO;
 
 /**
  * Created on 2017/4/18.
